+++ src/plone/volto/indexers.py
@@ -1,5 +1,7 @@
-from plone.volto.behaviors.preview import IPreview
+from Acquisition import aq_base
+from plone.dexterity.interfaces import IDexterityContent
 from plone.indexer.decorator import indexer
+from plone.volto.behaviors.preview import IPreview
 
 
 @indexer(IPreview)
@@ -11,3 +13,16 @@ def hasPreviewImage(obj):
     if obj.aq_base.preview_image:
         return True
     return False
+
+
+@indexer(IDexterityContent)
+def image_field_indexer(obj):
+    """Indexer for knowing in a catalog search if a content has any image."""
+    base_obj = aq_base(obj)
+
+    image_field = ""
+    if getattr(base_obj, "preview_image", False):
+        image_field = "preview_image"
+    elif getattr(base_obj, "image", False):
+        image_field = "image"
+    return image_field